refactor: modularize additional pages step
