+++ apps/cms/src/app/cms/wizard/steps/StepAdditionalPages/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./StepAdditionalPages";
